==> Notebook 5 XGBoost Parameter Tuning.ipynb <==
# Aujourd'hui on *s'amuse* avec le XGBoost


https://challengedata.ens.fr/en/challenge/39/prediction_of_transaction_claims_status.html

# Imports des librairies de bases

On ajoutera celles qui manquent au fur et à mesure de nos besoins

# To support both python 2 and python 3
from __future__ import division, print_function, unicode_literals

# Common imports
import numpy as np
import pandas as pd
import os, gc

# Définition de la seed pour le random

Très important pour qu'on voit les mêmes choses entre nos deux ordis

RANDOM_SEED = 42;
np.random.seed(RANDOM_SEED)

# Définition des paramètres pour Matplot

Rien de bien intéréssant

# To plot pretty figures
%matplotlib inline
import matplotlib
import matplotlib.pyplot as plt
plt.rcParams['axes.labelsize'] = 14
plt.rcParams['xtick.labelsize'] = 12
plt.rcParams['ytick.labelsize'] = 12

# Set des variables globales

Attention, je n'utilise les variables globales pour la gestion des fichiers. Sinon, c'est mort

# Where to save the figures
PROJECT_ROOT_DIR = "."
DATA_PROCESSED = os.path.join(PROJECT_ROOT_DIR, "data_processed")

# Fonction pour load les libraires

En vrai, on a juste besoin de pd.read_csv, mais c'était pour faire joli

def load_data(file,data_path=DATA_PROCESSED, sep=';'):
    csv_path = os.path.join(data_path, file)
    return pd.read_csv(csv_path, sep=';')

# On load les jeux de données

TX_data = load_data(file = "train.csv");

TX_data.info() # 42 colonnes, c'est un nombre qui fait plaisir. 
# [Note, il y en a plus maintenant, mais ce commentaire me fait toujours rire]

### Split entre train et test

from sklearn.model_selection import train_test_split

train_set, test_set = train_test_split(TX_data, 
                                       test_size=0.2, 
                                       random_state=RANDOM_SEED, 
                                       stratify=TX_data["CLAIM_TYPE"]
                                      )
del TX_data;

# Jointure entre les X et Y

from sklearn.preprocessing import Imputer

def datapreprocess(data):
    data=data.apply(pd.to_numeric, errors='ignore')
    
    # Y and X
    try :
        Y=data["CLAIM_TYPE"]
        X=data.drop("CLAIM_TYPE", axis=1,inplace=False)
    except:
        Y=0
        X=data
    # Exclude Objets
    X=X.select_dtypes(exclude=['object']) # j'exclude les variables catégorielles que j'ai oublié
    
    # Work on fare
    imp = Imputer(missing_values='NaN',strategy='median', axis=1)
    X=pd.DataFrame(imp.fit_transform(X),columns=X.columns.values)
 
    return X, Y

X_train, Y_train = datapreprocess(train_set)
X_test, Y_test = datapreprocess(test_set)

gc.collect()

from sklearn.metrics import roc_auc_score
from sklearn.preprocessing import LabelBinarizer

def multiclass_roc_auc_score(truth, pred):
    lb = LabelBinarizer()
    lb.fit(truth)
    return roc_auc_score(lb.transform(truth), lb.transform(pred), average="weighted")

from sklearn.metrics import confusion_matrix
def plot_confusion_matrix(matrix):
    """If you prefer color and a colorbar"""
    fig = plt.figure(figsize=(8,8))
    ax = fig.add_subplot(111)
    cax = ax.matshow(matrix)
    fig.colorbar(cax)

# MODEL!

from sklearn.utils.class_weight import compute_sample_weight

sample_weight_arr = compute_sample_weight(class_weight='balanced', y=Y_train)

## XGBoost

#### Core XGBoost Library VS scikit-learn API

Models can be trained in two different ways:

1. Directly using the core library – this is closer to the implementation of the caret-package in R
2. Using the scikit-learn API – this means that the models are implemented in a way that lets the scikit package recognize it as one of it’s own models.

Nous, on va travailler avec l'API de Sklearn, c'est pas optimisé mais plus simple. De toute façon, j'arrive pas à utiliser le Core, a cause des DMatrix qui veulent que des numerics en entrées

Doc des paramètres: https://github.com/dmlc/xgboost/blob/master/doc/parameter.md

Doc sur le tunning : https://www.analyticsvidhya.com/blog/2016/03/complete-guide-parameter-tuning-xgboost-with-codes-python/

from xgboost import XGBClassifier

###  XGBoost solo

params_XGB={
# General Parameters -  the overall functioning
    'booster':'gbtree',
    'silent':0,
    'nthread':4,
    'n_estimators' : 500,
    
# Booster Parameters - the individual booster (tree/regression) at each step
    'learning_rate' : 0.02,
    'min_child_weight' : 1, #A smaller value is chosen because it is a highly imbalanced class problem and leaf nodes can have smaller size groups.
    'max_depth' : 4,
    #'max_leaf_nodes':None, #If this is defined, GBM will ignore max_depth.
    'gamma' : 0.4,
    'max_delta_step':1, #it might help in logistic regression when class is extremely imbalanced/ 1-10 might help control the update
    'subsample' : 0.8,
    'colsample_bytree' : 0.9,
    'colsample_bylevel':1, #default
    'reg_lambda' : 1, #default
    'reg_alpha':0.05,
    'scale_pos_weight' : sample_weight_arr,

# Learning Task Parameters -  the optimization performed
    #'objective' : ['rank:pairwise','multi:softmax'], # you also need to set an additional num_class (number of classes)
    'objective' :['reg:gamma', 'rank:pairwise','multi:softmax'],
    'num_class' : len(Y_train.unique()),
    'eval_metric':["mlogloss", 'auc'],
    'seed' : RANDOM_SEED,
    'random_state':RANDOM_SEED
}

'base_score':0.5,
'booster':'gbtree',
'colsample_bylevel':1,
'colsample_bytree':0.8,
'eval_metric':'auc',
'gamma':0.3,
'learning_rate':0.09,
'max_delta_step':1,
'max_depth':2,
'min_child_weight':1,
'missing':None,
'n_estimators':100,
'n_jobs':1,
'nthread':None,
'num_class':8,
'objective':'multi:softprob',
'random_state':0,
'reg_alpha':0,
'reg_lambda':1,
'seed':42,
'silent':0,
'subsample':0.7

xgb_clf = XGBClassifier(**params_XGB)

eval_set = [(X_train, Y_train), (X_test, Y_test)]

params_XGB_fit={
    'X':X_train, 
    'y':Y_train, 
    'sample_weight':sample_weight_arr, 
    'eval_set':None, 
    'eval_metric':["mlogloss", 'auc'], 
    'early_stopping_rounds':None, 
    'verbose':False, 
    'xgb_model':None
}

xgb_clf.fit(**params_XGB_fit)

y_pred_xgb_train = xgb_clf.predict(X_train)
y_pred_xgb = xgb_clf.predict(X_test)

train_mAUC = multiclass_roc_auc_score(Y_train, y_pred_xgb_train)
test_mAUC = multiclass_roc_auc_score(Y_test, y_pred_xgb)

print("Performance sur le train : {}".format(train_mAUC))
print("Performance sur le test : {}".format(test_mAUC))

Performance sur le train : 0.7066348172305776

Performance sur le test : 0.6256314390125582

conf_mx = confusion_matrix(Y_test, y_pred_xgb)
row_sums = conf_mx.sum(axis=1, keepdims=True)
norm_conf_mx = conf_mx / row_sums
plot_confusion_matrix(norm_conf_mx)

#### C'est un beau score pour le XBoost

Cependant, j'ai optimisé pour la mauvaise métrique, et j'ai toujours pas fait le Grid Search

pd.DataFrame(xgb_clf.feature_importances_, index=X_train.columns, columns=["Feature"]).sort_values(by="Feature", ascending=False)

### XGBoost Grid Search Iterate

Comme mon PC est pourri, je vais chercher les paramètres de façon iterative

from sklearn.model_selection import GridSearchCV

#### Paramètres pour le XGB qui ne changent pas

params_XGB={
# General Parameters -  the overall functioning
    'booster':'gbtree',
    'silent':0,
    'nthread':4, # Je vais le commenter, puisque il détecte automatiquement le nombre de cores qu'il peut utiliser.
    'n_estimators' : 100,
    
# Booster Parameters - the individual booster (tree/regression) at each step
    'learning_rate' : 0.09,
    'colsample_bylevel':1, #default
    'reg_lambda' : 1, #default
    'scale_pos_weight' : sample_weight_arr,

# Learning Task Parameters -  the optimization performed
    'objective' : ['reg:gamma', 'rank:pairwise','multi:softmax'],
    'num_class' : len(Y_train.unique()),
    'eval_metric':["mlogloss", 'auc'],
    'seed' : RANDOM_SEED,
    'random_state':RANDOM_SEED
}

#### Paramètres pour la méthode `fit` de XGB qui ne changent pas

fit_params_xgb_cv={
    'sample_weight': sample_weight_arr, 
    'eval_set' : None, 
    'eval_metric' : ["mlogloss", 'auc'], 
    'early_stopping_rounds' : None, 
    'verbose':False, 
    'xgb_model':None
}

###  Optim 1 : max_depth, min_child_weight et max_delta_step

# Paramètres pour le GridSearch
params_XGB_CV = {
    'max_depth':range(2,5),
    'min_child_weight':range(1,6),
    'max_delta_step':list(range(1,5)),
}

xgb_gs_cv = GridSearchCV(XGBClassifier(**params_XGB), 
                              params_XGB_CV,
                              n_jobs=-1,
                              verbose=1)

xgb_gs_cv.fit(
    X = X_train, 
    y=Y_train, 
    groups=None, 
    **fit_params_xgb_cv
)
print(xgb_gs_cv.best_estimator_)
print("ROC score : {}".format(multiclass_roc_auc_score(Y_test, xgb_gs_cv.predict(X_test))))

Donc les paramètres optimaux sont:
1. `max_depth` : `4`
2. `min_child_weight` : `5`
3. `max_delta_step` : `4`

###  Optim 2 : gamma, et subsample

params_XGB={
# General Parameters -  the overall functioning
    'booster':'gbtree',
    'silent':0,
    'nthread':4, # Je vais le commenter, puisque il détecte automatiquement le nombre de cores qu'il peut utiliser.
    'n_estimators' : 100,
    'max_depth' : 4,
    'min_child_weight' : 5,
    'max_delta_step' : 4,
    
# Booster Parameters - the individual booster (tree/regression) at each step
    'learning_rate' : 0.09,
    'colsample_bylevel':1, #default
    'reg_lambda' : 1, #default
    'scale_pos_weight' : sample_weight_arr,

# Learning Task Parameters -  the optimization performed
    'objective' : ['reg:gamma', 'rank:pairwise','multi:softmax'],
    'num_class' : len(Y_train.unique()),
    'eval_metric':["mlogloss", 'auc'],
    'seed' : RANDOM_SEED,
    'random_state':RANDOM_SEED
}

# Paramètres pour le GridSearch
params_XGB_CV = {
    'gamma':[i/10.0 for i in range(0,5)],
    'subsample':[i/10.0 for i in range(6,10)],
}

xgb_gs_cv = GridSearchCV(XGBClassifier(**params_XGB), 
                              params_XGB_CV,
                              n_jobs=-1,
                              verbose=1)

xgb_gs_cv.fit(
    X = X_train, 
    y=Y_train, 
    groups=None, 
    **fit_params_xgb_cv
)
print(xgb_gs_cv.best_estimator_)
print("ROC score : {}".format(multiclass_roc_auc_score(Y_test, xgb_gs_cv.predict(X_test))))

Donc les paramètres optimaux sont:
1. `gamma` : `0.4`
2. `subsample` : `0.8`

###  Optim 3 : colsample_bytree, et reg_alpha

params_XGB={
# General Parameters -  the overall functioning
    'booster':'gbtree',
    'silent':0,
    'nthread':4, # Je vais le commenter, puisque il détecte automatiquement le nombre de cores qu'il peut utiliser.
    'n_estimators' : 100,
    'max_depth' : 4,
    'min_child_weight' : 5,
    'max_delta_step' : 4,
    'subsample' : 0.8,
    
# Booster Parameters - the individual booster (tree/regression) at each step
    'learning_rate' : 0.09,
    'colsample_bylevel':1, #default
    'reg_lambda' : 1, #default
    'scale_pos_weight' : sample_weight_arr,
    'gamma' : 0.4,

# Learning Task Parameters -  the optimization performed
    'objective' : ['reg:gamma', 'rank:pairwise','multi:softmax'],
    'num_class' : len(Y_train.unique()),
    'eval_metric':["mlogloss", 'auc'],
    'seed' : RANDOM_SEED,
    'random_state':RANDOM_SEED
}

# Paramètres pour le GridSearch
params_XGB_CV = {
    'colsample_bytree':[i/10.0 for i in range(6,10)],
    'reg_alpha':[0, 0.001, 0.005, 0.01, 0.05]
}

xgb_gs_cv = GridSearchCV(XGBClassifier(**params_XGB), 
                              params_XGB_CV,
                              n_jobs=-1,
                              verbose=1)

xgb_gs_cv.fit(
    X = X_train, 
    y=Y_train, 
    groups=None, 
    **fit_params_xgb_cv
)
print(xgb_gs_cv.best_estimator_)
print("ROC score : {}".format(multiclass_roc_auc_score(Y_test, xgb_gs_cv.predict(X_test))))

Donc les paramètres optimaux sont:
1. `colsample_bytree` : `0.9`
2. `reg_alpha` : `0.05`

###  Optim 4 : Le Learning rate 

Objectif : faire diminuer le risque d'overfit'

params_XGB={
# General Parameters -  the overall functioning
    'booster':'gbtree',
    'silent':0,
    'nthread':4, # Je le commente, puisque il détecte automatiquement le nombre de cores qu'il peut utiliser.
    'n_estimators' : 100,
    
# Booster Parameters - the individual booster (tree/regression) at each step
    #'learning_rate' : 0.09,
    'min_child_weight' : 5, #A smaller value is chosen because it is a highly imbalanced class problem and leaf nodes can have smaller size groups.
    'max_depth' : 4,
    #'max_leaf_nodes':None, #If this is defined, GBM will ignore max_depth.
    'gamma' : 0.4,
    'max_delta_step':4, #it might help in logistic regression when class is extremely imbalanced/ 1-10 might help control the update
    'subsample' : 0.8,
    'colsample_bytree' : 0.9,
    'colsample_bylevel':1, #default
    'reg_lambda' : 1, #default
    'reg_alpha':0.05,
    'scale_pos_weight' : sample_weight_arr,

# Learning Task Parameters -  the optimization performed
    'objective' : ['reg:gamma', 'rank:pairwise','multi:softmax'], # you also need to set an additional num_class (number of classes)
    'num_class' : len(Y_train.unique()),
    'eval_metric':["mlogloss", 'auc'],
    'seed' : RANDOM_SEED,
    'random_state':RANDOM_SEED
}

# Paramètres pour le GridSearch
params_XGB_CV = {
    'learning_rate':[i/1000.0 for i in range(3,22)],
}

xgb_gs_cv = GridSearchCV(XGBClassifier(**params_XGB), 
                              params_XGB_CV,
                              n_jobs=-1,
                              verbose=1)

xgb_gs_cv.fit(
    X = X_train, 
    y=Y_train, 
    groups=None, 
    **fit_params_xgb_cv
)
print(xgb_gs_cv.best_estimator_)
print("ROC score : {}".format(multiclass_roc_auc_score(Y_test, xgb_gs_cv.predict(X_test))))

###  Optim 5 : L'ensemble des paramètres trouvés

L'idée ici est de vérifier si les paramètres, ensemble, fonctionnent bien, donc on prend les paramètres trouvés, et on cherche un peu à gauche et à droite de chacun

params_XGB={
    #'base_score':0.5,
    'booster':'gbtree',
    'eval_metric':["mlogloss", 'auc'],
    'learning_rate':0.09,
    'missing':None,
    'n_estimators':100,
    'n_jobs':4,
    'nthread':4,
    'num_class':8,
    'objective':['reg:gamma', 'rank:pairwise','multi:softmax'],
    'random_state':RANDOM_SEED,
    'reg_lambda':1,
    'seed':RANDOM_SEED,
    'silent':0,
    
    'reg_alpha' : 0.05
    
}

params_XGB_CV={
    'colsample_bylevel':[1, 2], #
    'colsample_bytree':[0.8, 0.9, 1.0], #
    'gamma':[0.3, 0.4, 0.4], #
    'max_delta_step':[3,4], #
    'max_depth':[3,4], #
    'min_child_weight':[4,5], #
    'subsample':[0.6,0.7,0.8] #
}

xgb_gs_cv = GridSearchCV(XGBClassifier(**params_XGB), 
                              params_XGB_CV,
                              n_jobs=-1,
                              verbose=1)

xgb_gs_cv.fit(
    X = X_train, 
    y=Y_train, 
    groups=None, 
    **fit_params_xgb_cv
)
print(xgb_gs_cv.best_estimator_)
print("ROC score : {}".format(multiclass_roc_auc_score(Y_test, xgb_gs_cv.predict(X_test))))

conf_mx = confusion_matrix(Y_test, y_pred_xgb_cv)
row_sums = conf_mx.sum(axis=1, keepdims=True)
norm_conf_mx = conf_mx / row_sums
plot_confusion_matrix(norm_conf_mx)
